refactor(docs): register remember on the toolkit and derive the display from its key
the remember tool moves onto the docs toolkit with a browser execute, so the demo no longer uses the deprecated frontend registration; the component keeps the instruction block and the renderer.

the context display derives its whole context value from the incoming reset key rather than only its visibility, so a thread switch cannot show the previous thread's figure for a frame.
